// app.dezcablez/src/app/store/interfaces/item.ts
export interface IItem {
    id: string;
    name: string;
    description: string;
    type: string;
    price: number;
    stock: number;
    thumbnail: string;
    images?: string[];
    extra1Name?: string;
    extra1Value?: string;
    extra2Name?: string;
    extra2Value?: string;
    extra3Name?: string;
    extra3Value?: string;
    extra4Name?: string;
    extra4Value?: string;
    extra5Name?: string;
    extra5Value?: string;
    extra6Name?: string;
    extra6Value?: string;
    extra7Name?: string;
    extra7Value?: string;
  }